+++ src/classes/implements/Command/children/quarter.cpp
@@ -11,7 +11,7 @@ void Quarter::executeAction(CandyGame& Game){
             cout << "\nOops, kartu ability quarter-mu telah dimatikan sebelumnya :(.\nSilahkan lakukan perintah lain.\n" << endl;
         } else {
             cout << endl << playernow.getNickname() << " melakukan QUARTER! Poin hadiah";
-            int prevPoint = Game.getPoint();
+            long long int prevPoint = Game.getPoint();
             Game.setPoint(prevPoint*0.25);
             if(Game.getPoint() == 0){
                 Game.setPoint(1);
